refac: decouple api interface frm parkitright form
Previously the UI's form used the same fields as park it right. This was
nice because the API could simply pass through the fields to selenium,
but this resulted in a tight coupling of all of the services. Instead
this will decouple them into a custom (and cleaner IMO) format.